FIX: MySQL/MariaDB support packages in Ubuntu 18.04
This commit fixes the packages being installed in Ubuntu 18.04 to
support Ansible interaction with MySQL/MariaDB databases, as they
are not the same as Ubuntu 20.04/Debian Stable.

This was done by setting up a way to override only the `rails_app_mysql2`
variable in Ubuntu 18.xx and using the rest of the Debian environment.

The above is accomplished by passing a regex of files that are to be
included by the `include_var` task and takes advantage of the fact the
task loads the files in alphabetic order, thus in this case the file
for Ubuntu is loaded after the one for Debian.