FIX: Don't start application if it is not checkout
This commit fixes a failure when running the playbook and not
specifying a `rails_app_repo_url` and thus no checkout would
be done. This results in the failure to (re)start the application
at the end of the playbook.

This is the use case where the application will be deployed at a
later time, e.g. using Capistrano.

The fix is not to (re)start the application if the `rails_app_repo_url`
variable is not defined.