FIX: Use gem for database and `puma` gem from application Gemfile.
This commit ensures that the gems used for the database connection are
either already present in the `Gemfile` or are placed in the `Gemfile.local`.

The application service is then executed within the bundle for the application
to ensure that the proper versions are used.

This change is required for applications that define what versions of the
database gem and `puma` gem are to be used directly in their `Gemfile`.