CHG: Postgresql role and database creation tweaks.
This commit changes the way the postgresql role for the rails_app is
created and also the corresponding database.

The `rails_app` role is now created first and given the 'CREATEDB'
privilege, but dropping the 'SUPERUSER' privilege by using `NOSUPERUSER`.
This change is require to allow the user to destroy/create a database
from a `dump` file (e.g. initial load of database values).

This commit also changes the way the database is created to that
the `rails_app` role is the de facto owner of the database.