FIX: Minimal gem installation.
This commit changes the way the minimal gems `rake` and `bundler` are
installed because using the `gem` module can cause the gems not to
be installed for the rails_app user if the same/a version already
exists in the system.

The system installed gems seem not respect the bundler configuration for
the rails app being deployed thus causing trouble when deploying.